+++ src/main/java/ecdar/presentations/CanvasPresentation.java
@@ -1,16 +1,25 @@
 package ecdar.presentations;
 
+import com.jfoenix.controls.JFXRippler;
+import com.jfoenix.skins.ValidationPane;
 import ecdar.controllers.CanvasController;
 import ecdar.controllers.EcdarController;
+import ecdar.utility.colors.Color;
 import ecdar.utility.helpers.CanvasDragHelper;
 import ecdar.utility.helpers.MouseTrackable;
+import ecdar.utility.helpers.SelectHelper;
 import ecdar.utility.mouse.MouseTracker;
 import javafx.beans.property.BooleanProperty;
 import javafx.beans.property.DoubleProperty;
 import javafx.beans.property.SimpleBooleanProperty;
 import javafx.beans.property.SimpleDoubleProperty;
+import javafx.geometry.Bounds;
+import javafx.scene.Node;
+import javafx.scene.Parent;
 import javafx.scene.input.MouseButton;
+import javafx.scene.input.MouseEvent;
 import javafx.scene.layout.*;
+import javafx.scene.shape.Rectangle;
 
 public class CanvasPresentation extends Pane implements MouseTrackable {
     public MouseTracker mouseTracker;
@@ -34,6 +43,7 @@ public class CanvasPresentation extends Pane implements MouseTrackable {
         initializeGrid();
 
         CanvasDragHelper.makeDraggable(this, mouseEvent -> mouseEvent.getButton().equals(MouseButton.SECONDARY));
+        mouseTracker.registerOnMousePressedEventHandler(this::startDragBox);
 
         controller.zoomHelper.setGrid(this.grid);
         controller.zoomHelper.setCanvas(this);
@@ -119,5 +129,89 @@ public class CanvasPresentation extends Pane implements MouseTrackable {
 
     public CanvasController getController() {
         return controller;
+    }
+
+    private void startDragBox(final MouseEvent event) {
+        if(event.isPrimaryButtonDown()) {
+            SelectHelper.clearSelectedElements();
+
+            Rectangle selectionRectangle = new Rectangle();
+            selectionRectangle.setStroke(SelectHelper.SELECT_COLOR.getColor(SelectHelper.SELECT_COLOR_INTENSITY_BORDER));
+            javafx.scene.paint.Color fillColor = SelectHelper.SELECT_COLOR.getColor(Color.Intensity.I100);
+            fillColor = fillColor.deriveColor(fillColor.getHue(), fillColor.getSaturation(), fillColor.getBrightness(), 0.5);
+            selectionRectangle.setFill(fillColor);
+            selectionRectangle.getStrokeDashArray().addAll(5.0, 5.0);
+
+            double mouseDownX = event.getX();
+            double mouseDownY = event.getY();
+            selectionRectangle.setX(mouseDownX);
+            selectionRectangle.setY(mouseDownY);
+            selectionRectangle.setWidth(0);
+            selectionRectangle.setHeight(0);
+
+            mouseTracker.registerOnMouseDraggedEventHandler(e -> {
+                selectionRectangle.setX(Math.min(e.getX(), mouseDownX));
+                selectionRectangle.setWidth(Math.abs(e.getX() - mouseDownX));
+                selectionRectangle.setY(Math.min(e.getY(), mouseDownY));
+                selectionRectangle.setHeight(Math.abs(e.getY() - mouseDownY));
+
+                SelectHelper.clearSelectedElements();
+
+                // Recursively find all relevant ItemSelectable Nodes in the scene
+                updateSelection(this, selectionRectangle);
+            });
+
+            mouseTracker.registerOnMouseReleasedEventHandler(e -> this.getChildren().remove(selectionRectangle));
+            this.getChildren().add(selectionRectangle);
+        }
+    }
+
+    private void updateSelection(Parent currentNode, Rectangle selectionRectangle) {
+        // None of these nodes contain ItemSelectable nodes, so avoiding traversing these sub-trees improves performance
+        if (currentNode instanceof VBox || currentNode instanceof ValidationPane || currentNode instanceof JFXRippler || currentNode instanceof BorderPane) {
+            return;
+        }
+
+        if (currentNode instanceof LocationPresentation) {
+            if (selectIfWithinSelectionBox(currentNode, selectionRectangle)) {
+                SelectHelper.addToSelection(((LocationPresentation) currentNode).getController());
+            }
+        } else if (currentNode instanceof EdgePresentation) {
+            ((EdgePresentation) currentNode).getController().getNailNailPresentationMap().values().forEach((nailPresentation -> {
+                if (selectIfWithinSelectionBox(nailPresentation, selectionRectangle)) {
+                    SelectHelper.addToSelection(nailPresentation.getController());
+                }
+            }));
+        } else if (currentNode instanceof ComponentOperatorPresentation || currentNode instanceof ComponentInstancePresentation) {
+            if (selectIfWithinSelectionBox(currentNode, selectionRectangle)) {
+                SelectHelper.addToSelection((SelectHelper.ItemSelectable) currentNode);
+            }
+        }
+
+        currentNode.getChildrenUnmodifiable().forEach((node) -> {
+            if (node instanceof Parent) {
+                updateSelection((Parent) node, selectionRectangle);
+            }
+        });
+    }
+
+    /**
+     * Returns whether the item is within the selection.
+     * @param item the node to check.
+     * @param selectionRectangle the selection box.
+     */
+    private boolean selectIfWithinSelectionBox(Node item, Rectangle selectionRectangle) {
+        Bounds itemCoordinates = item.localToScreen(item.getLayoutBounds());
+        Bounds selectionCoordinates = selectionRectangle.localToScreen(selectionRectangle.getLayoutBounds());
+
+        if (selectionCoordinates == null || itemCoordinates == null) return false;
+
+        return selectionCoordinates.getMinX() <
+                itemCoordinates.getMinX() &&
+                itemCoordinates.getMinX() + 10 <
+                        selectionRectangle.getWidth() +
+                                selectionCoordinates.getMinX() &&
+                selectionCoordinates.getMinY() < itemCoordinates.getMinY() &&
+                itemCoordinates.getMinY() + 10 < selectionRectangle.getHeight() + selectionCoordinates.getMinY();
     }
 }
