+++ src/main/resources/ecdar/presentations/CanvasShellPresentation.fxml
@@ -10,7 +10,7 @@
       type="StackPane"
       fx:id="root"
       fx:controller="ecdar.controllers.CanvasShellController">
-    <CanvasPresentation fx:id="canvasPresentation" onMousePressed="#startDragBox"/>
+    <CanvasPresentation fx:id="canvasPresentation"/>
     <StackPane fx:id="toolbar"
                StackPane.alignment="TOP_LEFT"
                minHeight="56" maxHeight="56" maxWidth="224">
